+++ scratch.js
@@ -19,9 +19,17 @@ import nlp from './src/three.js'
 // console.log(three.pointer)
 // console.log(m.match('three').pointer)
 
-let doc = nlp('one match two three. four five match six')
-let res = doc.not('match two').debug()
-// res.debug()
+/*
+[_,a,a,a,-,-]
+[-,-,b,b,b,-]
+[-,x,x,x,x,-]
+*/
+
+let str = `one two three`
+let m = nlp(str).match('.')
+m = m.splitAfter('two')
+m.debug()
+// 'two three four'
 
 // we are hiring
 // we begin hiring